Refactor: Re-apply changes from PR #25 (Remove 'aioha' from codebase)
This commit re-applies the changes originally introduced in PR #25, which were aimed at removing the term 'aioha' and its variants from the entire codebase and replacing it with 'HiveFlutterKit' or its appropriate forms. These changes were inadvertently lost from the main branch due to a `git pull --no-rebase`.

Key changes re-applied include:

1.  **Project-wide Renaming:**
    - `aioha_flutter_core` to `hive_flutter_kit`
    - `AiohaFlutterCorePlugin` to `HiveFlutterKitPlugin`
    - `AiohaFlutterCoreWeb` to `HiveFlutterKitWeb`
    - `aioha_flutter_core_example` to `hive_flutter_kit_example`
    - Associated package names (e.g., `com.example.aioha_flutter_core` to `com.example.hive_flutter_kit`) and directory structures for Kotlin files were updated.
    - References in `README.md`, `example/README.md`, `.vscode/launch.json`, Android build files (`build.gradle.kts`, `AndroidManifest.xml`, `settings.gradle`), iOS files (`Info.plist`, test files), and web example files (`index.html`, `manifest.json`) have been updated.

2.  **`pubspec.yaml` Updates:**
    - Project `description` updated to reflect HiveFlutterKit.
    - `version` updated to `1.1.5` (as per PR #25).
    - Plugin registration class names and filenames updated.
    - Asset path for web JavaScript file updated from `web/aioha.js` to `web/hiveflutterkit.js`.

3.  **Website Homepage Features:**
    - Confirmed that `website/src/components/HomepageFeatures/index.js` uses PNG images (`undraw_docusaurus_mountain.png`, etc.) with `<img>` tags and `useBaseUrl`, consistent with the intent of PR #25 to use PNGs for these features.

This commit restores the comprehensive renaming and refactoring efforts of PR #25.
