+++ README.md
@@ -4,16 +4,6 @@
 
 ---
 
-# Aioha
-
-## All-In-One Hive Authentication
-
-### Introduction
-
-- Aioha (All-In-One Hive Authentication) - Now HiveFlutterKit is an API that provides a common interface for working with different Hive login providers.
-- This allows easier integration of Hive login and transacting on the network with fewer code.
-- For Javascript package usage, https://aioha.dev/docs
-
 ### HiveFlutterKit
 
 - HiveFlutterKit is a Flutter package that provides a common interface for working with different Hive login providers.
