+++ android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'aioha_flutter_core'
+rootProject.name = 'hive_flutter_kit'
