+++ android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.example.aioha_flutter_core">
+  package="com.example.hive_flutter_kit">
 </manifest>
