+++ example/android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <application
-        android:label="aioha_flutter_core_example"
+        android:label="hive_flutter_kit_example"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
