+++ example/android/app/src/main/kotlin/com/example/hive_flutter_kit_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.aioha_flutter_core_example
+package com.example.hive_flutter_kit_example
 
 import io.flutter.embedding.android.FlutterActivity
