+++ example/web/index.html
@@ -18,19 +18,19 @@
 
   <meta charset="UTF-8">
   <meta content="IE=Edge" http-equiv="X-UA-Compatible">
-  <meta name="description" content="Demonstrates how to use the aioha_flutter_core plugin.">
+  <meta name="description" content="Demonstrates how to use the Hive Flutter Kit plugin.">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
   <!-- iOS meta tags & icons -->
   <meta name="mobile-web-app-capable" content="yes">
   <meta name="apple-mobile-web-app-status-bar-style" content="black">
-  <meta name="apple-mobile-web-app-title" content="aioha_flutter_core_example">
+  <meta name="apple-mobile-web-app-title" content="hive_flutter_kit_example">
   <link rel="apple-touch-icon" href="icons/Icon-192.png">
 
   <!-- Favicon -->
   <link rel="icon" type="image/png" href="favicon.png"/>
 
-  <title>aioha_flutter_core_example</title>
+  <title>Hive Flutter Kit Example</title>
   <link rel="manifest" href="manifest.json">
 </head>
 <body>
