+++ example/web/manifest.json
@@ -1,11 +1,11 @@
 {
-    "name": "aioha_flutter_core_example",
-    "short_name": "aioha_flutter_core_example",
+    "name": "hive_flutter_kit_example",
+    "short_name": "hive_flutter_kit_example",
     "start_url": ".",
     "display": "standalone",
     "background_color": "#0175C2",
     "theme_color": "#0175C2",
-    "description": "Demonstrates how to use the aioha_flutter_core plugin.",
+    "description": "Demonstrates how to use the hive_flutter_kit plugin.",
     "orientation": "portrait-primary",
     "prefer_related_applications": false,
     "icons": [
